add cursorimg.TestCursors and clean up TestHotspot
